+++ nens_auth_client/requests_session.py
@@ -5,16 +5,16 @@ from .jwt_utils import decode_jwt
 import time
 
 
-def get_session_for(remote_user: RemoteUser, valid_seconds=60):
-    """Return a requests.Session enriched with a OAuth2 access token configured.
+def get_session_for(remote_user: RemoteUser, valid_seconds: int = 60):
+    """Return a requests.Session enriched with OAuth2 access tokens configured.
 
     Automatically refreshes if the access token is out of date; in that case, the RemoteUser
     will be updated with a new access_token and refresh_token.
 
     Args:
         remote_user: the RemoteUser to get the tokens from
-        valid_secons: if the access_token is valid less than this many seconds, then
-            the token will be refreshed.
+        valid_seconds: if the access_token is valid less than this many seconds then
+            the token will be refreshed
 
     Returns:
         requests.Session subclass (authlib.integrations.requests_client.OAuth2Session)
@@ -32,7 +32,7 @@ def get_session_for(remote_user: RemoteUser, valid_seconds=60):
     )
 
     # Note: Authlib / requests doesn't expose a means to refresh a token when
-    # necessary. We do it here with a
+    # necessary. We do it here with a single check against this server's time.
     if (session.token.expires_at - valid_seconds) < int(time.time()):
         session.refresh_token(client.load_server_metadata()["token_endpoint"])
         remote_user.id_token = session.token.id_token